feat: add tc-003
refactored existing code-base to have a bridgenode builder
tc-002 nits in compositions and renaming
created a local tc-003 composition to test out locally
