+++ compositions/cluster-k8s/tc-002-da-sync/40-40-20-100/320-100mib-200ms.toml
@@ -1,10 +1,10 @@
 [metadata]
-  name = "tc-002-da-sync-40-40-20-100-set"
+  name = "002-da-sync-40-40-20-100-set"
   author = "Bidon15"
 
 [global]
   plan = "celestia"
-  case = "tc-002-da-sync"
+  case = "002-da-sync"
   total_instances = 200
   builder = "docker:generic"
   runner = "cluster:k8s"
@@ -12,7 +12,7 @@
 
 [global.run.test_params]
   persistent-peers = "10"
-  submit-times = "10"
+  submit-times = "12"
   msg-size = "100000"
   validator = "40"
   bridge = "40"
@@ -59,7 +59,7 @@
     [groups.run.test_params]
     latency = "200"
     bandwidth = "320Mib"
-    block-height = "13"
+    block-height = "11"
     role = "bridge"
 
 [[groups]]
@@ -81,7 +81,7 @@
     [groups.run.test_params]
     latency = "200"
     bandwidth = "320Mib"
-    block-height = "12"
+    block-height = "10"
     role = "full"
 
 [[groups]]
@@ -103,5 +103,5 @@
     [groups.run.test_params]
     latency = "200"
     bandwidth = "100Mib"
-    block-height = "12"
+    block-height = "10"
     role = "light"
