+++ manifest.toml
@@ -28,16 +28,33 @@ instances = { min = 1, max = 100, default = 3 }
     msg-size = { type = "int", default = 10000}
 
 [[testcases]]
-name = "tc-002-da-sync"
+name = "002-da-sync"
 instances = { min = 4, max = 3000, default = 12 }
     [testcases.params]
     latency = { type = "int", default = 0}
     bandwidth = { type = "string", default = "256Mib"}
     validator = { type = "int", default = 3}
     persistent-peers = { type = "int", default = 3}
+    submit-times = { type = "int", default = 4}
+    msg-size = { type = "int", default = 10000}
     bridge = { type = "int", default = 3}
     full = { type = "int", default = 3}
     light = { type = "int", default = 3}
     block-height = { type = "int" }
     role = { type = "string" }
 
+[[testcases]]
+name = "003-full-sync-past"
+instances = { min = 4, max = 3000, default = 12 }
+    [testcases.params]
+    latency = { type = "int", default = 0}
+    bandwidth = { type = "string", default = "256Mib"}
+    validator = { type = "int", default = 3}
+    persistent-peers = { type = "int", default = 3}
+    submit-times = { type = "int", default = 4}
+    msg-size = { type = "int", default = 10000}
+    bridge = { type = "int", default = 3}
+    full = { type = "int", default = 3}
+    light = { type = "int", default = 3}
+    block-height = { type = "int" }
+    role = { type = "string" }
