+++ tests/tc-003-full-sync-past.go
@@ -2,21 +2,22 @@ package tests
 
 import (
 	nodesync "github.com/celestiaorg/test-infra/tests/node-sync"
+	syncpast "github.com/celestiaorg/test-infra/tests/sync-past"
 
 	"github.com/testground/sdk-go/run"
 	"github.com/testground/sdk-go/runtime"
 )
 
-// Test-Case #002 - DA nodes are in sync with validators
+// Test-Case #003 - Full nodes are syncing past headers faster then validators produce new ones
 // Description is in docs/test-plans/001-Big-Blocks/test-cases
-func SyncNodes(runenv *runtime.RunEnv, initCtx *run.InitContext) (err error) {
+func FullSyncPast(runenv *runtime.RunEnv, initCtx *run.InitContext) (err error) {
 	switch runenv.StringParam("role") {
 	case "validator":
 		err = nodesync.RunAppValidator(runenv, initCtx)
 	case "bridge":
-		err = nodesync.RunBridgeNode(runenv, initCtx)
+		err = syncpast.RunBridgeNode(runenv, initCtx)
 	case "full":
-		err = nodesync.RunFullNode(runenv, initCtx)
+		err = syncpast.RunFullNode(runenv, initCtx)
 	case "light":
 		err = nodesync.RunLightNode(runenv, initCtx)
 	}
